producer class -> single method
