--- python-client/onestop_client/tests/primitive_string.avsc
@@ -1,3 +1,0 @@
-{
-    "type": "string"
-}
